Update README.md
Final Project
Developing an Interactive System with the MSP432
The project was to design and implement an interactive embedded system based on the MSP432 with an LCD display, keypad for user input, and peripheral devices to be controlled such
as LEDs, a DC motor, a servo motor that would represent opening a 'door', as well as interrupt buttons to disable/enable the LEDs at their last PWM setting and to disable the DC motor.
The LCD display was dimmable via a potentiometer to an ADC enabled pin on the MSP432. See the final project report for a full detail of the functionality.